refactoring/term(#37): Création de la classe Term
