+++ src/Service/CategoryService.php
@@ -1,15 +1,16 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Service;
 
 use App\Entity\Category;
-use Framework\Database\EntityManagerInterface;
 use Framework\Utils\Slugger;
+use Framework\Database\EntityManagerInterface;
 
 class CategoryService
 {
-    public function __construct(private EntityManagerInterface $em)
-    {}
+    public function __construct(private EntityManagerInterface $em) {}
 
     public function create(Category $category): void
     {
